src/requirements.txt
aiogram==2.17.1
aiohttp==3.8.1
aioredis==2.0.0
aiosignal==1.2.0
async-timeout==4.0.1
attrs==21.2.0
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.9
emoji==1.6.1
flatbuffers==2.0
frozenlist==1.2.0
gunicorn==20.1.0
idna==3.3
lxml==4.7.1
multidict==5.2.0
numpy==1.21.4
onnxruntime-gpu==1.10.0
protobuf==3.19.1
pytz==2021.3
sentencepiece==0.1.96
typing-extensions==4.0.1
uvloop==0.16.0
yarl==1.7.2
